# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EntreSkill Hub API
  description: >
    Skill-to-Startup Enablement Platform API. Covers auth, assessment,
    recommendations, roadmaps, learning resources, mentors, notifications,
    and admin operations. See docs/API.md for narrative documentation.
  version: "1.0.0"
  contact:
    name: EntreSkill Hub Engineering
servers:
  - url: https://entreskillhub.com/api
    description: Production
  - url: https://staging.entreskillhub.com/api
    description: Staging
  - url: http://localhost:3000/api
    description: Local development
...
paths:
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Always returns 200 regardless of whether the email exists
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using a signed token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "200": { description: Password updated }
        "400": { description: Invalid or expired token }
...
  /assessment:
    patch:
      tags: [Assessment]
      summary: Submit or update skill/interest profiling questionnaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentInput"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationError"
  /recommendations:
    get:
      tags: [Recommendations]
      summary: Get ranked business idea recommendations for the current user
      parameters:
        - name: refresh
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/RecommendedIdea" }
                  fallback: { type: boolean }
...
  /bookmarks:
    get:
      tags: [Bookmarks]
      summary: List current user's bookmarks
      responses:
        "200": { description: OK }
    post:
      tags: [Bookmarks]
      summary: Toggle a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetType, targetId]
              properties:
                targetType: { type: string, enum: [idea, resource] }
                targetId: { type: string }
      responses:
        "200": { description: New bookmark state }
  /mentors:
    get:
      tags: [Mentors]
      summary: List verified mentors
      security: []
      parameters:
        - name: expertise
          in: query
          schema: { type: string }
        - name: availabilityDay
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MentorProfile" }
  /mentors/apply:
    post:
      tags: [Mentors]
      summary: Apply to become a mentor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MentorApplication" }
      responses:
        "201": { description: Mentor application created (pending verification) }
  /mentors/{id}/request-session:
    post:
      tags: [Sessions]
      summary: Request a mentor session
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
      responses:
        "201": { description: Session request created }
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List/filter users (admin only)
      parameters:
        - name: role
          in: query
          schema: { type: string, enum: [user, mentor, admin] }
        - name: verificationStatus
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200": { description: OK }
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Forbidden:
      description: Authenticated but role/ownership does not permit this action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
    User:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [user, mentor, admin] }
        skills: { type: array, items: { type: string } }
        interests: { type: array, items: { type: string } }
        location:
          type: object
          properties:
            state: { type: string }
            district: { type: string }
            isRural: { type: boolean }
    AssessmentInput:
      type: object
      required: [skills, interests, location]
      properties:
        skills: { type: array, items: { type: string } }
        interests: { type: array, items: { type: string } }
        experienceLevels:
          type: object
          additionalProperties: { type: string, enum: [beginner, intermediate, advanced] }
        location:
          type: object
          properties:
            state: { type: string }
            district: { type: string }
            isRural: { type: boolean }
        timeCommitment: { type: string, enum: [part-time, full-time] }
    BusinessIdea:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        description: { type: string }
        relatedSkills: { type: array, items: { type: string } }
        difficultyLevel: { type: string, enum: [beginner, intermediate, advanced] }
        estimatedStartupCostINR:
          type: object
          properties:
            min: { type: number }
            max: { type: number }
        roadmapId: { type: string }
        isActive: { type: boolean }
...
    RecommendedIdea:
      allOf:
        - $ref: "#/components/schemas/BusinessIdea"
        - type: object
          properties:
            score: { type: number }
            reasons: { type: array, items: { type: string } }
...
    MentorProfile:
      type: object
      properties:
        _id: { type: string }
        userId: { type: string }
        expertiseAreas: { type: array, items: { type: string } }
        yearsExperience: { type: integer }
        bio: { type: string }
        verificationStatus: { type: string, enum: [pending, verified, rejected] }
        availability:
          type: array
          items:
            type: object
            properties:
              day: { type: string }
              startTime: { type: string }
              endTime: { type: string }
    MentorApplication:
      type: object
      required: [bio, expertiseAreas, yearsExperience]
      properties:
        bio: { type: string }
        expertiseAreas: { type: array, items: { type: string } }
        yearsExperience: { type: integer }
        supportingLinks: { type: array, items: { type: string } }
security:
  - sessionCookie: []